requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
beautifulsoup4
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub




